deface-refactor/requirements_complete.txt
# Core dependencies
paramiko>=2.12.0
redis>=5.0.0
rq>=1.16.1
rich>=13.7.0

# FastAPI and server
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
email-validator>=2.0.0

# Additional utilities
python-dotenv>=1.0.0